rm HadoopLayerUpdater
