+++ spark/src/test/scala/geotrellis/spark/io/hadoop/HadoopSpaceTimeSpec.scala
@@ -16,7 +16,6 @@ abstract class HadoopSpaceTimeSpec
   type Container = RasterRDD[SpaceTimeKey]
 
   lazy val reader = HadoopLayerReader[SpaceTimeKey, Tile, RasterRDD](outputLocal)
-  lazy val updater = HadoopLayerUpdater[SpaceTimeKey, Tile, RasterRDD](outputLocal)
   lazy val tiles = HadoopTileReader[SpaceTimeKey, Tile](outputLocal)
   lazy val sample =  CoordinateSpaceTime
 }
