+++ spark/src/test/scala/geotrellis/spark/io/hadoop/HadoopSpatialSpec.scala
@@ -13,7 +13,6 @@ abstract class HadoopSpatialSpec
           with AllOnesTestTileTests {
   type Container = RasterRDD[SpatialKey]
   lazy val reader = HadoopLayerReader[SpatialKey, Tile, RasterRDD](outputLocal)
-  lazy val updater = HadoopLayerUpdater[SpatialKey, Tile, RasterRDD](outputLocal)
   lazy val tiles = HadoopTileReader[SpatialKey, Tile](outputLocal)
   lazy val sample = AllOnesTestFile
 }
